% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date_column.R
\name{add_date_column}
\alias{add_date_column}
\title{Function adds a column of class \code{Date} to a \code{data.frame}}
\usage{
add_date_column(dt, time_columns, column_name)
}
\arguments{
\item{dt}{The target \code{data.frame} or \code{data.table} for which a column of class \code{Date} will be appended.}

\item{time_columns}{A vector of column names from \code{dt} that have time related numeric values. The names should
be listed in the year:month:day:seconds order.}

\item{column_name}{A string that sets the name of the new \code{Date} column.}
}
\value{
A new \code{data.table} with an added \code{Date} column
}
\description{
Given a vector of the submitted \verb{data.frame's} column names that have time related values (such as
year, month, day etc.), the function will add a column to the submitted \code{data.frame}
that has a class of \code{Date}. The function uses the package function \code{anytime::anydate()} to
compute the \code{Date} values.
}
\author{
Rick Dean
}
