% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_xts.R
\name{df_to_xts}
\alias{df_to_xts}
\title{Function converts a data.frame to an xts time object}
\usage{
df_to_xts(df, time_col, data_cols)
}
\arguments{
\item{df}{A data.frame containing time and data columns}

\item{time_col}{A string that is the name for the time column from \code{df}.}

\item{data_cols}{A string or vector string containing the name(s) of data columns from \code{df}.}
}
\value{
An \code{xts} time series object
}
\description{
The data.frame must have a time based Date or POSIXct column along
with one or more data columns. Function returns an \code{xts} time object.
}
\author{
Rick Dean
}
