% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Date.R
\name{get_Date}
\alias{get_Date}
\title{Function creates Date time object(s) for representing calendar dates}
\usage{
get_Date(date_vals = NULL, date_format = NULL)
}
\arguments{
\item{date_vals}{A character string or vector of character
strings of time in the assumed ISO 8601 format of "YYYY-mm-dd".}

\item{date_format}{A character string that defines the format of
\code{date_vals} if different from the ISO 8601 format. The format
string should follow the format string for \code{base::strptime()}.}
}
\value{
A Date object(s)
}
\description{
The function is just a simple wrapper around the
\code{base::as.Date()} as a reminder of the arguments required
to implement the base function. Provides a list of "try out"
formats.
}
\author{
Rick Dean
}
