% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_POSIXct.R
\name{get_POSIXct}
\alias{get_POSIXct}
\title{Function creates a POSIXct time object(s) for representing calendar
dates and times}
\usage{
get_POSIXct(time_vals = NULL, time_format = NULL, time_origin = NULL)
}
\arguments{
\item{time_vals}{A vector/object as character in date-time format or
a numeric in seconds.}

\item{time_format}{A character string that defines the format of
\code{time_vals} if different from the ISO 8601 format. The format
string should follow the format string for \code{base::strptime()}.}

\item{time_origin}{A character string that sets the origin/start date
for calculating the internal elapsed seconds representation of the time.
The default in R is January 1, 1970.}
}
\value{
A POSIXct time object
}
\description{
The function is just a simple wrapper around the
\code{base::as.POSIXct()} as a reminder of the arguments required
to implement the base function.
}
\author{
Rick Dean
}
