% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series_data.R
\name{get_series_data}
\alias{get_series_data}
\title{Retrieve the observations of a ts/zoo/xts time series object}
\usage{
get_series_data(series = NULL)
}
\arguments{
\item{series}{Either a \code{ts},\code{zoo} or \code{xts} time object.}
}
\value{
A numeric/character vector of observations
}
\description{
Function will retrieve the data of a
\code{ts},\code{zoo} or \code{xts} object.
}
\author{
Rick Dean
}
