% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SP500_symbols.R
\name{get_SP500_symbols}
\alias{get_SP500_symbols}
\title{Get stock market symbols for companies in the S&P 500 Index}
\usage{
get_SP500_symbols()
}
\value{
A data frame with variables for S&P 500 symbols, company names, and company sectors.
}
\description{
Returns a character vector of S&P 500 symbols using Wikipedia as a source.
}
\author{
Rick Dean
}
