% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiingo_dividends.R
\name{get_tiingo_dividends}
\alias{get_tiingo_dividends}
\title{Get stock market dividends via Tiingo.}
\usage{
get_tiingo_dividends(
  symbols = "IBM",
  periodicity = "monthly",
  from = "2019-01-01",
  to = "2019-12-31"
)
}
\arguments{
\item{symbols}{A string of stock market symbols to be accessed.}

\item{periodicity}{A string that defines the frequency of sampled prices. Acceptable values are \dQuote{daily},
\dQuote{monthly}, \dQuote{quarterly}, \dQuote{yearly}.}

\item{from}{A string or Date object that sets the starting date in CCYY-MM-DD format.}

\item{to}{A string or Date object that sets the ending date in CCYY-MM-DD format.}
}
\value{
A data frame with \dQuote{Date} and \dQuote{DivCash} along with other stock price related variables.
}
\description{
A wrapper around the riingo::riingo_prices()) for returning a data frame of historic stock market dividends.
For further information on setting this package's arguments see \href{https://cran.r-project.org/web/packages/riingo/riingo.pdf}{riingo}.
Note that you need to set a Tinngo API key before calling this function. See the above reference for further
information. See \code{RtsaPkg::read_tingo_key()} if you have saved the key at the .Renviron file.
}
\author{
Rick Dean
}
