% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiingo_stock_prices.R
\name{get_tiingo_stock_prices}
\alias{get_tiingo_stock_prices}
\title{Get stock market prices via Tiingo}
\usage{
get_tiingo_stock_prices(
  symbols = "IBM",
  periodicity = "weekly",
  from = "2019-01-01",
  to = "2019-12-31"
)
}
\arguments{
\item{symbols}{A character vector of stock market symbols to be accessed.}

\item{periodicity}{A string that defines the frequency of sampled prices. Acceptable values are "daily",
"weekly", "monthly", "annually".}

\item{from}{A string or Date object that sets the starting date in CCYY-MM-DD format.}

\item{to}{A string or Date object that sets the ending date in CCYY-MM-DD format.}
}
\value{
A data frame with historic "Date"'s along with stock price related variables.
}
\description{
A wrapper around the riingo::riingo_prices()) for returning a data frame of historic stock prices.
For further information on setting this package's arguments see \href{https://cran.r-project.org/web/packages/riingo/riingo.pdf}{riingo}.
Note that you need to set a Tinngo API key before calling this function. See the above reference for further
information. See \code{RtsaPkg::read_tingo_key()} if you have saved the key at the .Renviron file.
}
\author{
Rick Dean
}
