% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yahoo_stock_prices.R
\name{get_yahoo_stock_prices}
\alias{get_yahoo_stock_prices}
\title{Get stock market prices via Yahoo}
\usage{
get_yahoo_stock_prices(
  symbols = "GOOG",
  from = "2019-01-01",
  to = "2019-12-31",
  periodicity = "daily",
  return_class = "data.frame"
)
}
\arguments{
\item{symbols}{A character vector of stock market symbols to be accessed.}

\item{from}{A string or Date object that sets the starting date in CCYY-MM-DD format.}

\item{to}{A string or Date object that sets the ending date in CCYY-MM-DD format.}

\item{periodicity}{A string that defines the sampling of the data. Accepted values are \dQuote{daily}, \dQuote{weekly},
\dQuote{monthly}.}

\item{return_class}{A string that sets the class of the return object. Acceptable values are \code{data.frame} or \code{xts}.}
}
\value{
A data frame with a \code{date} column or xts object along with values for Open, High, Low, Close, Volume, Adjusted.
}
\description{
Function is a wrapper around the quantmod package for returning a data frame or xts object of historic stock market prices.
Function provides access to various stock market parameters via the getSymbols() of quantmod.  Yahoo is the source for the data.
For further information on setting this package's arguments see \href{https://cran.r-project.org/web/packages/quantmod/}{quantmod}
}
\author{
Rick Dean
}
