% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_decompose.R
\name{graph_decompose}
\alias{graph_decompose}
\title{Function plots an observed time series along with its decomposed seasonal, trend, and irregular components}
\usage{
graph_decompose(
  series_ts = NULL,
  type_comp = "additive",
  title = NULL,
  subtitle = NULL,
  x_title = "DateTime",
  y_title = "Value",
  x_limits = NULL,
  x_major_breaks = waiver(),
  x_major_date_breaks = waiver(),
  x_date_labels = waiver(),
  show_major_grids = TRUE,
  show_minor_grids = TRUE,
  col_width = 10,
  row_height = 2.4,
  silent_NA_warning = FALSE
)
}
\arguments{
\item{series_ts}{A time series (ts) object.}

\item{type_comp}{A string that describes the type of seasonal component. Accepted values are
\dQuote{additive} and \dQuote{multiplicative}.}

\item{title}{A string that defines an overall title to the multi-paneled plot.}

\item{subtitle}{A string that defines an overall subtitle.}

\item{x_title}{A string that defines the x axis title.}

\item{y_title}{A string that defines the y axis title.}

\item{x_limits}{A Date/POSIXct 2 element vector that sets the minimum and maximum for the x axis.
Use NA to refer to the existing minimum and maximum.}

\item{x_major_breaks}{A Date/POSIXct vector or function that defines the exact major tic locations along the x axis.}

\item{x_major_date_breaks}{For Date/POSIXct, a string containing the number and date unit for major breaks.
Examples: \code{"1 year"}, \code{"4 sec"}, \code{"3 month"}, \code{"2 week"}.}

\item{x_date_labels}{For Date/POSIXct, a string containing the format codes for the x axis date format.
This can be a strftime format for each x axis tic date label.
Examples: \code{"\%Y-\%m"}, \code{"\%Y/\%b/\%d"}, \code{"\%H-\%M-\%S"}.}

\item{show_major_grids}{A logical that controls the appearance of major grids.}

\item{show_minor_grids}{A logical that controls the appearance of minor grids.}

\item{col_width}{An integer that sets the width of each plot column in inches.}

\item{row_height}{An integer that sets the height of each plot column in inches.}

\item{silent_NA_warning}{A logical that controls the appearance of a console warning when Na's
are removed.}
}
\value{
Returns a named list with:
\enumerate{
\item \dQuote{decompose_df} -- A data.frame and data.table with column values for the
observed series, time, seasonal, trend, and random.
\item \dQuote{plots} -- A multi-paneled TableGrob object containing line plots of the
series' observed values, trend, seasonal, and random values across time.
Use \code{grid::grid.draw(plots)} to display the plots.
}
}
\description{
Function uses \code{decompose} from the stats package to plot the series' observed values, seasonal,
trend, and irregular components using moving averages. User can set the seasonal component as additive or multiplicative.

Function uses the \href{https://github.com/deandevl/RplotterPkg}{RplotterPkg} package to draw the multi-paneled plot.
}
\author{
Rick Dean
}
