% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_dif.R
\name{graph_dif}
\alias{graph_dif}
\title{Function focuses on calculating and plotting the lagged difference across a time series.}
\usage{
graph_dif(
  df = NULL,
  time_col = NULL,
  value_col = NULL,
  dif_lag = 1,
  title = NULL,
  x_title = NULL,
  y_title = NULL,
  x_limits = NULL,
  x_major_breaks = waiver(),
  x_major_date_breaks = waiver(),
  x_date_labels = waiver(),
  y_limits = NULL,
  y_major_breaks = waiver(),
  show_pts = TRUE,
  show_obs = TRUE,
  show_major_grids = TRUE,
  show_minor_grids = TRUE,
  col_width = 30,
  row_height = 8,
  display_plot = TRUE,
  png_file_path = NULL
)
}
\arguments{
\item{df}{A data.frame containing a time series with both a value and time column.}

\item{time_col}{Names the column from \code{df} for the time values. Values can
be numeric or Date/POSIXct.}

\item{value_col}{Names the numeric column from \code{df} for time series values.}

\item{dif_lag}{An integer that sets the lag of difference to use.}

\item{title}{A string that defines an overall title to the pair of plots.}

\item{x_title}{A string that sets the x axis title.}

\item{y_title}{A string that sets the y axis title.}

\item{x_limits}{A Date/POSIXct 2 element vector that sets the minimum and maximum for the x axis.
Use NA to refer to the existing minimum and maximum.}

\item{x_major_breaks}{A Date/POSIXct vector or function that defines the exact major tic locations along the x axis.}

\item{x_major_date_breaks}{For Date/POSIXct, a string containing the number and date unit for major breaks.
Examples: \code{"1 year"}, \code{"4 sec"}, \code{"3 month"}, \code{"2 week"}.}

\item{x_date_labels}{For Date/POSIXct, a string containing the format codes for the x axis date format.
This can be a strftime format for each x axis tic date label.
Examples: \code{"\%Y-\%m"}, \code{"\%Y/\%b/\%d"}, \code{"\%H-\%M-\%S"}.}

\item{y_limits}{A numeric 2 element vector that sets the minimum and maximum for the difference y axis.
Use NA to refer to the existing minimum and maximum.}

\item{y_major_breaks}{A numeric vector or function that defines the exact major tic locations for the difference y axis'.}

\item{show_pts}{A logical which if FALSE will plot only the lines.}

\item{show_obs}{A logical which if FALSE hides the plot of observations.}

\item{show_major_grids}{A logical that controls the appearance of major grids.}

\item{show_minor_grids}{A logical that controls the appearance of minor grids.}

\item{col_width}{An integer that sets the width of each plot column in centimeters.}

\item{row_height}{An integer that sets the height of each plot column in centimeters.}

\item{display_plot}{A logical that if TRUE displays the plot.}

\item{png_file_path}{A character string with the directory and file name to produce
a png image of the plot.}
}
\value{
Returning a named list with:
\enumerate{
\item "diff_df" -- A data frame with column variables for time "DateTime" and
the difference values "Value".
\item "plots" -- A multi-paneled TableGrob object plotting the differences and optionally
the observed series. Use \code{grid::grid.draw(plots)} to display the plots.
}
}
\description{
The lagged differences across a time series at a certain degree \code{dif_lag}
is taken primarily to transform the series from a non-stationary to stationary process.
Function returns a multi-paneled plot of the observed time series (optional), along with a
plot for the series' lagged differences. The actual difference values are also returned
in a data.frame.
}
\details{
The function uses the \href{https://github.com/deandevl/RplotterPkg}{RplotterPkg::multi_panel_grid}
package to draw the multi-paneled plot.

If \code{display_plot} is TRUE then the plots will be displayed. If \code{display_plot} is FALSE then
the function returns a named list that includes a plot object which can be displayed from the console by entering:
\enumerate{
\item \code{grid::grid.newpage()}
\item \code{grid::grid.draw(plot object)}
}
}
\author{
Rick Dean
}
