% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_ma.R
\name{graph_ma}
\alias{graph_ma}
\title{Function focuses on calculating and plotting the moving average of an observed time series.}
\usage{
graph_ma(
  df = NULL,
  time_col = NULL,
  value_col = NULL,
  window_n = 4,
  ma_type = "sma",
  overlap = TRUE,
  title = NULL,
  ma_caption = NULL,
  ob_caption = NULL,
  x_title = NULL,
  y_title = NULL,
  x_limits = NULL,
  x_major_breaks = waiver(),
  x_major_date_breaks = waiver(),
  x_date_labels = waiver(),
  y_limits = NULL,
  y_major_breaks = waiver(),
  show_major_grids = TRUE,
  show_minor_grids = TRUE,
  show_pts = TRUE,
  show_observe = TRUE,
  col_width = 18,
  row_height = 5,
  display_plot = TRUE,
  png_file_path = NULL
)
}
\arguments{
\item{df}{A data frame with variables for times and corresponding values.}

\item{time_col}{Names the column from \code{df} for the time values. Values can
be numeric or Date/POSIXct.}

\item{value_col}{Names the value column from \code{df}.}

\item{window_n}{An integer that controls the backward window length of the moving average.}

\item{ma_type}{A string that sets the type of moving average. Accepted values are "sma" simple, "tri" triangular,
"wma" weighted, "exp" exponential, "mod" modified, and "spe" Spencer weighted 15 point average.}

\item{overlap}{A logical which if \code{TRUE} overlaps both the observed and the moving average series'. If \code{FALSE} the
plots are in separate panels.}

\item{title}{A string that sets the overall title to the plots.}

\item{ma_caption}{A string that sets the caption for the ma plot.}

\item{ob_caption}{A string that sets the caption for the observed plot.}

\item{x_title}{A string that defines the x axis title.}

\item{y_title}{A string that defines the y axis title.}

\item{x_limits}{A Date/POSIXct 2 element vector that sets the minimum and maximum for the x axis.
Use NA to refer to the existing minimum and maximum.}

\item{x_major_breaks}{A Date/POSIXct vector or function that defines the exact major tic locations along the x axis.}

\item{x_major_date_breaks}{For Date/POSIXct, a string containing the number and date unit for major breaks.
Examples: \code{"1 year"}, \code{"4 sec"}, \code{"3 month"}, \code{"2 week"}.}

\item{x_date_labels}{For Date/POSIXct, a string containing the format codes for the x axis date format.
This can be a strftime format for each x axis tic date label.
Examples: \code{"\%Y-\%m"}, \code{"\%Y/\%b/\%d"}, \code{"\%H-\%M-\%S"}.}

\item{y_limits}{A numeric 2 element vector that sets the minimum and maximum for the y axis.
The default is \code{c(1,10)}.}

\item{y_major_breaks}{A numeric vector or function that defines the exact major tic locations for the moving average y axis'.}

\item{show_major_grids}{A logical that controls the appearance of major grids.}

\item{show_minor_grids}{A logical that controls the appearance of minor grids.}

\item{show_pts}{A logical which if FALSE will plot only the lines.}

\item{show_observe}{A logical that controls the appearance of the observed time series.}

\item{col_width}{An numeric that sets the width of each plot in centimeters.}

\item{row_height}{A numeric that sets the height of each plot in centimeters.}

\item{display_plot}{A logical that if TRUE displays the plot.}

\item{png_file_path}{A character string with the directory and file name to produce
a png image of the plot.}
}
\value{
Returning a named list with:
\enumerate{
\item "ma_df" -- A data.frame/data.table with column variables for time "DateTime" and the moving average values "Value" and
source of the values "Source".
\item "plots" -- A multi-panelled/overlapped TableGrob object plotting the observed series and the moving averages.
Use \code{grid::grid.draw(plots)} to display the plots.
}
}
\description{
Function returns either an overlapped or multi-paneled plot of the observed time series,
along with the plot for the moving average. The actual moving average values are also returned in a data.frame.
}
\details{
The function uses the \href{https://github.com/deandevl/RplotterPkg}{RplotterPkg::multi_panel_grid}
package to draw the multi-paneled plot.

If \code{display_plot} is TRUE then the plots will be displayed. If \code{display_plot} is FALSE then
the function returns a named list that includes a plot object which can be displayed from the console by entering:
\enumerate{
\item \code{grid::grid.newpage()}
\item \code{grid::grid.draw(plot object)}
}
}
\author{
Rick Dean
}
