% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_quandl_key.R
\name{read_quandl_key}
\alias{read_quandl_key}
\title{Read the Quandl api key for the Quandl package}
\usage{
read_quandl_key()
}
\value{
returns the api key string
}
\description{
Function reads the \code{QUANDL_KEY} token from the .Renviron file.

To edit the .Renviron file run \code{usethis::edit_r_environ()}.  An api key named QUANDL_KEY
was defined in this file with the user's Quandl auth key gotten from the Quandl web site.
}
\author{
Rick Dean
}
