% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_to_df.R
\name{ts_to_df}
\alias{ts_to_df}
\title{Convert a "ts" time series object to a "data.frame" object}
\usage{
ts_to_df(ts_obj = NULL, col_name = NULL)
}
\arguments{
\item{ts_obj}{Either a univariate time series object of class \dQuote{ts} or a multivariate time series object \dQuote{mts}.}

\item{col_name}{If \code{ts_obj} is of class \dQuote{ts} then this argument sets the returned data column name.
The default data column name is \dQuote{V1}.}
}
\value{
A data.frame/data.table with a column of time named (\dQuote{DateTime}) and the series' data column(s).
}
\description{
Function returns a data.frame/data.table with a Date/Time column named \dQuote{DateTime} and column(s) of
associated data.
}
\author{
Rick Dean
}
