% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsObj_to_dt.R
\name{tsObj_to_dt}
\alias{tsObj_to_dt}
\title{Function converts a time series object to a \code{data.table}}
\usage{
tsObj_to_dt(series = NULL, wide_form = FALSE)
}
\arguments{
\item{series}{A \code{ts}, \code{xts}, \code{mts}, or \code{zoo} time series object}

\item{wide_form}{A logical which if TRUE will reshape the data.table to wide form.}
}
\value{
A \code{data.table}
}
\description{
The function \code{tsbox::ts_df()} is incorporated to return a \code{data.table} from
a time object of classes \code{ts}, \code{xts}, \code{mts}, or \code{zoo}. In the \code{data.table} the series' time index
is named "time" and corresponding values as "value". If the series is a multiple time series (mts)
then an index column is created (named "id") with the names of the variables.
}
