% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xts_to_df.R
\name{xts_to_df}
\alias{xts_to_df}
\title{Convert an "xts" time series object to a data.frame/data.table object}
\usage{
xts_to_df(xts_obj = NULL, col_names = NULL, time_class = "Date")
}
\arguments{
\item{xts_obj}{Either a univariate or multivariate time series object of class \dQuote{xts}.}

\item{col_names}{A string vector from which to set the returned data.frame's column names.
The first column is a date or time column.}

\item{time_class}{A string that sets the class of the returned data.frame's time column.
Acceptable values are \dQuote{Date}, \dQuote{POSIXct}, or \dQuote{POSIXlt}}
}
\value{
A data.frame/data.table with a column of time named (\dQuote{index}) and the series' data column(s).
}
\description{
Function returns a data.frame/data.table with a Date/Time
column named \dQuote{index} and column(s) of associated data.
}
\author{
Rick Dean
}
