% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series_info.R
\name{get_series_info}
\alias{get_series_info}
\title{Function gets both time and starting data information on a \code{ts}, \code{zoo}, or
\code{xts} time series object}
\usage{
get_series_info(series = NULL, n_obser = NULL)
}
\arguments{
\item{series}{Either a  \code{ts}, \code{zoo}, or \code{xts} time series object}

\item{n_obser}{An integer that sets the number of observations to display. If NULL
then no observations are included in the data.frame.}
}
\value{
A named list with the series statistics(named "stats") and sample values(named "data").
}
\description{
The function returns a short synopsis data.frame of a time series object
and optionally its starting observations.
}
\author{
Rick Dean
}
