% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_decompose.R
\name{graph_decompose}
\alias{graph_decompose}
\title{Function plots an observed time series along with its decomposed seasonal, trend, and irregular components}
\usage{
graph_decompose(
  series_ts = NULL,
  type_comp = "additive",
  title = NULL,
  x_title = "DateTime",
  x_limits = NULL,
  x_major_breaks = waiver(),
  x_major_date_breaks = waiver(),
  x_date_labels = waiver(),
  col_width = 20,
  row_height = 5,
  display_plot = TRUE
)
}
\arguments{
\item{series_ts}{A time series (class ts) object.}

\item{type_comp}{A string that describes the type of seasonal component. Accepted values are
"additive" and "multiplicative".}

\item{title}{A string that defines an overall title to the multi-paneled plot.}

\item{x_title}{A string that defines the x axis title.}

\item{x_limits}{A Date/POSIXct 2 element vector that sets the minimum and maximum for the x axis.
Use NA to refer to the existing minimum and maximum.}

\item{x_major_breaks}{A Date/POSIXct vector or function that defines the exact major tic locations along the x axis.}

\item{x_major_date_breaks}{For Date/POSIXct, a string containing the number and date unit for major breaks.
Examples: \code{"1 year"}, \code{"4 sec"}, \code{"3 month"}, \code{"2 week"}.}

\item{x_date_labels}{For Date/POSIXct, a string containing the format codes for the x axis date format.
This can be a strftime format for each x axis tic date label.
Examples: \code{"\%Y-\%m"}, \code{"\%Y/\%b/\%d"}, \code{"\%H-\%M-\%S"}.}

\item{col_width}{An integer that sets the width of each plot column in centimeters.}

\item{row_height}{An integer that sets the height of each plot column in centimeters.}

\item{display_plot}{A logical that if TRUE displays the plot.}
}
\value{
Returns a named list with:
\enumerate{
\item "decompose_dt" -- A data.frame and data.table with column values for the
observed series, time, seasonal, trend, and random.
\item "plots" -- A multi-paneled TableGrob object containing line plots of the
series' observed values, trend, seasonal, and random values across time.
Use \code{grid::grid.draw(plots)} to display the plots.
}
}
\description{
Function uses \code{decompose} from the stats package to plot the series' observed values, seasonal,
trend, and irregular components using moving averages. User can set the seasonal component as additive or
multiplicative.
}
\details{
Function uses the \href{https://github.com/deandevl/RplotterPkg}{RplotterPkg::multi_panel_grid} package to
draw the multi-paneled plot.

if \code{display_plot} is TRUE then the plots will be displayed. If \code{display_plot} is FALSE then
the function returns a named list that includes a plot object which can be displayed from the console by entering:
\enumerate{
\item \code{grid::grid.newpage()}
\item \code{grid::grid.draw(plot object)}
}
}
\author{
Rick Dean
}
